pyproject.toml
[project]
name = "education-tutoring-system"
version = "0.1.0"
description = "LangGraph-based multi-agent education tutoring system"
requires-python = ">=3.10"
dependencies = [
    "langgraph>=0.2.0",
    "langgraph-checkpoint-sqlite>=1.0.0",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "openai>=1.0.0",
    "aiosqlite>=0.19.0",
    "pydantic>=2.0.0",
    "tenacity>=8.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "hypothesis>=6.0.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
]
api = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "**/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
